feat(main): JsonSer is somewhat working now (i.e. no errors)
Still has many issues. The TODO for the main one is copied below:

TODO: Polymorphism is LOST COMPLETELY here. This is a problem.
This *must* be fixed for CardEffect to arrive in an intelligible format!

How to fix: detect/code which dataclass has to have type/class info preserved.
Deser might not need to support this - we might not be sending cards the other way for now.
Anyway, loading arbitrary classes seems a bit dubious, security-wise.

Also, added a small example in main_v2.py of serialising the entire game state (pre-start, pre-deal cards)