fix(frontend): Ignore no-scopable-elements svelte CSS warning
This is required because ButtonDiv needs some way of receiveing styles and that way is currently class attribute + `:global`.
I (obviously) want these styles next to the each usage of ButtonDiv so I refuse to put it in an external CSS file.
(I'm using ButtonDiv as a custom element rather than a svelte-style component - and it doesn't seem to like that)